+++ sjf4j-core/src/incubator/java/org/sjf4j/mapper/package-info.java
@@ -1,5 +1,5 @@
 /**
- * Structural mapping APIs for converting one object graph or node model into another.
+ * Incubating structural mapping APIs for converting one object graph or node model into another.
  *
  * <p><b>Deprecated.</b> Use {@link org.sjf4j.annotation.mapping.CompiledMapper} instead.</p>
  */