+++ sjf4j-core/src/main/java/org/sjf4j/bytecode/BytecodePath.java
@@ -1,4 +1,4 @@
-package org.sjf4j.compiled;
+package org.sjf4j.bytecode;
 
 import org.sjf4j.TypeReference;