+++ sjf4j-testbench/src/test/java/org/sjf4j/testbench/asm/BytecodePathTest.java
@@ -2,7 +2,7 @@ package org.sjf4j.testbench.asm;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
-import org.sjf4j.compiled.BytecodePath;
+import org.sjf4j.bytecode.BytecodePath;
 import org.sjf4j.exception.JsonException;
 
 import java.util.ArrayList;